+++ dvh/modules/tools/utilities.py
@@ -82,8 +82,10 @@ def moving_avg_by_calendar_day(xyw, avg_days):
     return x_final, moving_aves
 
 
-def flatten_list_of_lists(some_list, remove_duplicates=False):
+def flatten_list_of_lists(some_list, remove_duplicates=False, sort=False):
     data = [item for sublist in some_list for item in sublist]
+    if sort:
+        data.sort()
     if remove_duplicates:
         return list(set(data))
     return data